requirements.txt
# Core NLP / HF stack
transformers
sentencepiece
datasets
evaluate

# Metrics
sacrebleu
rouge_score

# ML / DL backend (PyTorch only)
torch

# Utilities
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
notebook

# AWS
boto3
mypy-boto3-s3

# Config handling
python-box

# API layer (Python 3.12 compatible)
fastapi
uvicorn
Jinja2

# Critical compatibility pins
protobuf
ml-dtypes
